requirements.txt
PyQt6==6.6.1
python-nmap==0.7.1
reportlab==4.0.8
requests==2.31.0
python-dotenv==1.0.0
SQLAlchemy==2.0.25 